+++ bindings/ruby/lib/stateset_embedded.rb
@@ -7,5 +7,5 @@ rescue LoadError
 end
 
 module StateSet
-  VERSION = '0.7.5'
+  VERSION = '0.7.6'
 end