+++ cli/bin/stateset-update.js
@@ -41,7 +41,7 @@ OPTIONS:
 EXAMPLES:
   stateset update
   stateset update status --json
-  stateset update apply --tag 0.7.5 --yes
+  stateset update apply --tag 0.7.6 --yes
 `;
 
 function parseTimeout(value) {